Fix compiling on 64-bit architectures.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@128836 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sbin/ggate/ggatec/ggatec.c
@@ -254,7 +254,7 @@ once_again:
 			goto done;
 		}
 		if (ggio.gctl_cmd == BIO_READ) {
-			if (bsize < ggio.gctl_length) {
+			if (bsize < (size_t)ggio.gctl_length) {
 				ggio.gctl_data = realloc(ggio.gctl_data,
 				    ggio.gctl_length);
 				if (ggio.gctl_data != NULL)
